+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - Change Image(sub menu) functionality added for active apps
   - App overview 'in review' style fix
   - Enhance Sub Menu by adding 'Change Description' for active apps
+  - Change Description(sub menu) functionality added for active apps
 - Bugfix
   - Service Release Process
     - Service Release process not working